# Specialize Vec::from_elem to use calloc