@leonardo-m 2017-03-10
What's the performance for short (n <= 8)?