@mbrubeck 2017-03-10
For the following benchmarks, using the default allocator (alloc_jemalloc) on Linux:

```
#[bench]
fn bench_big(b: &mut test::Bencher) {
    b.iter(|| test::black_box(vec![0u8; 1024 * 1024]));
}
#[bench]
fn bench_medium(b: &mut test::Bencher) {
    b.iter(|| test::black_box(vec![0u8; 1024]));
}
#[bench]
fn bench_small(b: &mut test::Bencher) {
    b.iter(|| test::black_box(vec![0u8; 8]));
}
```

Before this PR:

```
test bench_big    ... bench:     231,338 ns/iter (+/- 97,246)
test bench_medium ... bench:          34 ns/iter (+/- 1)
test bench_small  ... bench:          24 ns/iter (+/- 3)
```

After this PR:

```
test bench_big    ... bench:         853 ns/iter (+/- 47)
test bench_medium ... bench:          37 ns/iter (+/- 2)
test bench_small  ... bench:          25 ns/iter (+/- 1)

```