@clarcharr 2017-03-10
Is there a reason why this can't be specialised for all integer types?