@mbrubeck 2017-03-10
> Is there a reason why this can't be specialised for all integer types if the value is zero?

It can, and I'm working on a patch to do that. It could also be specialized for floats, and for some less-common types like `Option<&T>` and raw pointers; I'm not sure where the point of diminishing returns is.