@clarcharr 2017-03-10
I think that specialising for integer zeroes is definitely reasonable (because `[0; n]` is the most common way of initialising an array). I personally would like to see the others you mentioned too, although that's just my opinion. :P