@solson 2017-03-14
@whitequark made a good point on IRC: If we had a `Pod` (plain old data) trait, you could make this specialization for all `T: Pod` in a single impl. This would handle many of the types mentioned above but it would also handle arbitrary user-defined types that meet the `Pod` requirements.

Previous discussions [here](https://github.com/rust-lang/rfcs/pull/1387) and [here](https://github.com/rust-lang/rfcs/pull/936).