@sfackler 2017-03-15
@solson how would that work here though? This only handles the case of an all-zero initial bit pattern.