@solson 2017-03-15
@sfackler For `T: Pod`, it's safe to use `std::mem::zeroed::<T>()`.