@clarcharr 2017-03-15
Also one thing to consider is that `Zeroable` is already part of libcore and could be useful here for this optimisation.