@bluss 2017-03-15
@solson the missing part of the explanation is how to identify that `x` is equivalent to (all bits) zero in `vec![x; n]` for `T: Pod`.