@mbrubeck 2017-03-15
Added a patch to specialize `vec![0; n]` for other integer types.