@solson 2017-03-15
@bluss One basic way would be to iterate over the bytes of `x` and check that they are all zero (which seems like something the optimizer could easily simplify). We could perhaps do something a bit more direct with `[0; size_of::<T>()]` in the future when `size_of` is a `const fn`.