@frewsxcv 2017-03-17
This fails to compile on Windows:

@bors r- https://ci.appveyor.com/project/rust-lang/rust/build/1.0.2427/job/7o459p1om5u0gus6