@mbrubeck 2017-03-17
Added a missing `unsafe` that broke the build on Windows.  It might be a good idea to `try` this PR before trying to land it though, since I still haven't tested the Windows code.