@bluss 2017-03-18
Just to link issues: Related to tracking issue #27389 (global, default allocator)