@arielb1 2017-03-19
Fix allocator-dummy.rs?

@bors r-